requirements.txt
Flask==3.0.0
line-bot-sdk==3.6.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
